+++ src/app/[locale]/layout.tsx
@@ -7,7 +7,6 @@ import React from "react";
 
 import { BackgroundFigures } from "@/src/components/backgroundImages/BackgroundFigures";
 import { BackgroundImages } from "@/src/components/backgroundImages/BackgroundImages";
-import { GreetingPortal } from "@/src/components/christmasHomePageDecoration/modal/GreetingPortal";
 import { CookiesComponent } from "@/src/components/cookies/Cookies";
 import { Footer } from "@/src/components/footer/Footer";
 import { Header } from "@/src/components/header/Header";
@@ -74,11 +73,11 @@ export async function generateMetadata({
 
 export default function LocaleLayout({
     children,
-    modal,
+    // modal,
     params: { locale },
 }: Readonly<{
     children: React.ReactNode;
-    modal: React.ReactNode;
+    // modal: React.ReactNode;
     params: { locale: string };
 }>) {
     const messages = useMessages();
@@ -100,12 +99,12 @@ export default function LocaleLayout({
                    dark:bg-purple-400 dark:text-grey bg-white-100 text-greyLight`}
                     >
                         <Providers>
-                            <GreetingPortal />
+                            {/* <GreetingPortal /> */}
                             <BackgroundImages />
                             <div className="min-h-screen flex flex-col">
                                 <Header />
                                 <main className="flex-auto min-h-full">
-                                    {modal}
+                                    {/* {modal} */}
                                     <div className="pt-[80px] pc:pt-[0px] pc:ml-[80px] deskxl:ml-[120px] ">
                                         <BackgroundFigures />
                                         {children}
